Remove old serializers and viewsets
